+++ packages/drizzle/src/factory/resolver.ts
@@ -51,6 +51,7 @@ import { GraphQLInt, GraphQLNonNull } from "graphql"
 import {
   type DrizzleResolverFactoryOptions,
   DrizzleWeaver,
+  type SelectiveTable,
   type TableSilk,
 } from ".."
 import { getSelectedColumns, inArrayMultiple } from "../helper"
@@ -71,6 +72,7 @@ import type {
   BaseDatabase,
   CountQuery,
   DeleteMutation,
+  DrizzleQueriesResolver,
   InferRelationTable,
   InferTableName,
   InferTableRelationalConfig,
@@ -530,6 +532,34 @@ export abstract class DrizzleResolverFactory<
         [`insertInto${capitalize(name)}Single`]: this.insertSingleMutation(),
         [`update${capitalize(name)}`]: this.updateMutation(),
         [`deleteFrom${capitalize(name)}`]: this.deleteMutation(),
+      },
+      options
+    ) as any
+  }
+
+  public queriesResolver<
+    TTableName extends string = TTable["_"]["name"],
+  >(options?: {
+    name?: TTableName
+    middlewares?: Middleware[]
+  }): ObjectChainResolver<
+    GraphQLSilk<SelectiveTable<TTable>, SelectiveTable<TTable>>,
+    DrizzleQueriesResolver<TTable, TTableName>
+  > {
+    const name = options?.name ?? this.tableName
+
+    const fields: Record<string, Loom.Field<any, any, any, any>> = mapValue(
+      this.db._.schema?.[this.tableName]?.relations ?? {},
+      (_, key) => this.relationField(key)
+    )
+
+    return loom.resolver.of(
+      this.output,
+      {
+        ...fields,
+        [name]: this.selectArrayQuery(),
+        [`${name}Single`]: this.selectSingleQuery(),
+        [`${name}Count`]: this.countQuery(),
       },
       options
     ) as any